requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1

# LLM & Embeddings
openai==1.30.1
sentence-transformers==2.7.0
tiktoken==0.7.0

# Vector Stores
faiss-cpu==1.8.0
chromadb==0.5.0

# Document Loaders
pypdf==4.2.0
markdown==3.6
beautifulsoup4==4.12.3
requests==2.32.2
html2text==2024.2.26

# Retrieval
rank-bm25==0.2.2
numpy==1.26.4
scipy==1.13.0

# Reranking
transformers==4.41.1
torch==2.3.0

# Evaluation
ragas==0.1.9
datasets==2.19.1
pandas==2.2.2

# Observability & Utilities
langsmith==0.1.63
python-dotenv==1.0.1
pyyaml==6.0.1
tenacity==8.3.0
rich==13.7.1
typer==0.12.3
loguru==0.7.2

# Caching
diskcache==5.6.3

# Testing
pytest==8.2.1
pytest-asyncio==0.23.7
httpx==0.27.0
pytest-cov==5.0.0
